+++ lec1_inheritance/Laptop.java
@@ -18,7 +18,7 @@ public class Laptop extends Computer {
     }
 
     //new constructor with parameters
-    public Laptop(int ram, int speed, int hDDspace, int batteryLeft)
+    public Laptop(long ram, int speed, int hDDspace, int batteryLeft)
     {
         super(ram, speed, hDDspace);
         this.batteryLeft = batteryLeft;
