+++ lec1_polymorphism/Laptop.java
@@ -20,7 +20,7 @@ public class Laptop extends Computer implements Operable{
     }
 
     //new constructor with parameters
-    public Laptop(int ram, int speed, int hDDspace, int batteryLeft)
+    public Laptop(long ram, int speed, int hDDspace, int batteryLeft)
     {
         super(ram, speed, hDDspace);
         this.batteryLeft = batteryLeft;
